+++ HISTORY.rst
@@ -2,6 +2,14 @@
 History
 =======
 
+0.2.0 (2019-12-12)
+---------------------
+
+* Default pvalue set to 1e-8
+
+* Added term_size to json output which contains number of genes associated with term
+
+
 0.1.0 (2019-09-19)
 ------------------